feat(mobile): personaliza a jornada e corrige o cálculo de séries efetivas
Séries efetivas (correção de domínio)
- A v1 exigia RIR registrado. Como RIR é opcional no contrato, quem não anota
  reserva — o usuário padrão — via "0 séries efetivas" em todos os grupos
  musculares: a métrica lia como "você não treinou".
- A v1 também contava série inteira ou nada, sem noção de ativação: o glúteo de
  quem só agacha aparecia com 0 séries efetivas ao lado de 25 t de carga.
- O que permitiu os dois defeitos passarem: o mock reimplementava a agregação
  inline em vez de chamar @atlas/domain, e as duas cópias divergiram.
- A v2 conta a série de trabalho a menos que haja evidência explícita de que foi
  leve, e pondera pelo activationWeight do próprio catálogo. Fracionário no
  contrato, arredondado na apresentação. effectiveSetsV1 fica para
  reprodutibilidade; CALCULATION_VERSION vai a 2. O texto do app que descrevia a
  fórmula antiga foi reescrito.

Personalização
- Contrato: BodyRegion (oito regiões articulares — sem diagnóstico),
  ExperienceLevel na mesma escala de Difficulty, TrainingPreferences no perfil
  (null = não respondeu, diferente de [] = nenhuma) e stressedRegions no
  exercício.
- Os 60 exercícios do fixture foram etiquetados com critério discriminante, não
  exaustivo: marcar "ombro" nos vinte movimentos de membro superior devolveria
  lista vazia a quem tem ombro sensível. Um teste exige que excluir qualquer
  região deixe mais da metade do catálogo de pé.
- exercise-fit.ts decide adequação com motivos em código: nada some em silêncio,
  limitação não é diagnóstico, e nível filtra para cima mas nunca esconde —
  esconder o agachamento de um iniciante é dizer que ele nunca vai agachar.
- Onboarding ganhou o passo "Suas preferências" e passou a perguntar sexo
  biológico, que já existia no contrato e nunca era coletado; por isso a tela de
  metas mostrava "faltam dados" no primeiro dia de uso.
- Catálogo filtra com interruptor visível e explica o que saiu; o montador de
  ficha ordena mas não esconde, porque ali o usuário pode legitimamente querer o
  exercício que marcou para poupar.

Foto de perfil no lugar do avatar vetorial
- Saíram avatar-parts/presets/assets, AvatarEditor, AvatarScreen, AvatarSwatch,
  Avatar, a rota /avatar/edit e o entitlement premiumAvatarItems. AvatarConfig
  (nove campos e três cores) virou photoUri.
- O construtor custava um catálogo de geometria, uma tela, um item de paywall e
  uma etapa na jornada para entregar um boneco que ninguém confunde com a
  própria cara. PersonAvatar já caía para as iniciais, então o fallback existia.

Heros sem corte vertical
- Os assets são retrato (1080x1440 e 960x1440) e viviam numa caixa de 212 pt: na
  largura da tela isso recortava mais da metade da altura e decepava a figura
  central. A altura passou a vir da proporção medida de cada arte.

typecheck, lint, vitest (277, +25), check:parity e expo export web verdes.
Onboarding percorrido na prévia web a 414 px.

+++ atlas/packages/contracts/src/schemas/insights.ts
@@ -7,7 +7,14 @@ export const MuscleVolume = z.object({
   muscleCode: z.string(),
   displayName: z.string(),
   weightedVolumeKg: z.number().min(0),
-  effectiveSets: z.number().int().min(0),
+  /**
+   * Fracionário desde a v2 do cálculo: um agachamento vale uma série efetiva
+   * para o quadríceps e cerca de meia para o glúteo. Inteiro obrigava a
+   * escolher entre 1 e 0 para o sinergista, e a escolha era 0 — o que fazia o
+   * glúteo de quem só agacha aparecer com "0 séries efetivas" ao lado de 25 t
+   * de carga. Ver `weightedEffectiveSets` em @atlas/domain. A UI arredonda.
+   */
+  effectiveSets: z.number().min(0),
   /** 0..1, normalizado contra o grupo de maior volume no período. */
   intensity: z.number().min(0).max(1),
 });
